+++ app/scripts/components/dynamic_tabs/graph_tabs.js
@@ -5,7 +5,7 @@ export const graphTabsComponent = () => ({
     template: html`
         <uib-tab ng-repeat="data in $ctrl.tabs" graph-ctrl="graph-ctrl">
             <uib-tab-heading ng-class="{not_loading: progress > 99}"
-                >{{'graph' | loc:lang}}
+                >{{'graph' | loc:$root.lang}}
                 <div class="tab_progress" style="width:{{progress || 0}}%" ng-show="loading"></div>
                 <span ng-click="closeTab($index, $event)" tab-spinner="tab-spinner"></span>
             </uib-tab-heading>