+++ app/scripts/components/korp_error.js
@@ -9,7 +9,7 @@ export const korpErrorComponent = {
             <object class="korp_fail inline-block" type="image/svg+xml" data="${korpFailImg}">
                 <img class="korp_fail" src="${korpFailImg}" />
             </object>
-            <div class="fail_text inline-block">{{'fail_text' | loc:lang}}</div>
+            <div class="fail_text inline-block">{{'fail_text' | loc:$root.lang}}</div>
         </div>
     `,
 }