+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 
 - Disable Trend diagram and Map buttons while waiting for statistics search to finish [#346](https://github.com/spraakbanken/korp-frontend/issues/346)
+- Fix dynamic translation for tabs etc
 
 ### Removed